requirements.txt
pygame-ce>=2.1.4
pygame-gui>=0.6.9